+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.107.0] - Not released
+### Removed
+- Do not try to preview shorten TikTok URLs in posts.
 
 ## [1.106.1] - 2022-02-03
 ### Fixed