+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update "privacy" cookie if it is already set. It should prevent 2-weeks cookie
   expiration in Safari.
 
+### Added
+- New frontend setting: comments.hideRepliesToBanned (false by default, can be
+  adjusted on Settings / Appearance page). If it is set to true, comments that
+  starts with a @-mention of banned user are replaced by placeholder ('Comment
+  with reply to blocked user'). Comments authored by the current user are always
+  visible.
+
 ## [1.106.1] - 2022-02-03
 ### Fixed
 - Re-release. 1.106.0 has incorrect merge.